Backend/service6/requirements.txt
# Base Django & REST Framework
Django>=4.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.2.0

# Excel support
pandas>=2.2.2
openpyxl>=3.1.2
django-import-export>=3.3.6

# Communication interservices
py_eureka_client>=0.11.0
requests>=2.31.0
python-decouple>=3.8

# PDF & Impression
weasyprint>=61.0
reportlab>=4.0.8
PyPDF2>=3.0.1

# Cloudinary (si tu gères les fichiers distants)
cloudinary>=1.39.0
mysqlclient>=2.2.0


# Parsing XML si besoin
xmltodict>=0.13.0
